fakeapiモジュールを追加し、fake関連の処理を移動する
