+++ .idea/gradle.xml
@@ -13,6 +13,7 @@
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/api" />
             <option value="$PROJECT_DIR$/app" />
+            <option value="$PROJECT_DIR$/fakeapi" />
           </set>
         </option>
         <option name="resolveModulePerSourceSet" value="false" />
