+++ fakeapi/build.gradle.kts
@@ -1,0 +1,24 @@
+plugins {
+    kotlin("multiplatform")
+    id("com.android.library")
+    id("org.jetbrains.kotlin.plugin.serialization")
+}
+
+kotlin {
+    android()
+
+    sourceSets {
+        val commonMain by getting {
+            dependencies {
+                implementation(project(":api"))
+                implementation("io.ktor:ktor-client-serialization:1.6.8")
+                api("io.ktor:ktor-client-mock:1.6.8")
+            }
+        }
+        val commonTest by getting {
+            dependencies {
+                implementation(kotlin("test"))
+            }
+        }
+    }
+}
