+++ settings.gradle
@@ -16,3 +16,4 @@ dependencyResolutionManagement {
 rootProject.name = "KMMSample"
 include ':app'
 include ':api'
+include ':fakeapi'
